+++ ChoGGi's Library/Changelog.md
@@ -3,6 +3,7 @@
 ## v3.0 (Unreleased)
 - Returns names of (some) tables.
 - Shows function file/line number instead of "function: 0000000008D7C100".
+- Converted UI to Use TextStyle.
 
 ## v2.9 (30 Oct 2018)
 - AnimDebug_Toggle wasn't removing oriention obj.